Removed infinite loop (stack overflow) when creating delete confirmation

+++ PastLoot/Core/PastLoot.lua
@@ -5,21 +5,8 @@ local L = LibStub("AceLocale-3.0"):GetLocale("PastLoot")
 local LDB = LibStub:GetLibrary("LibDataBroker-1.1")
 -- local LDBIcon = LibStub("LibDBIcon-1.0")
 
--- used to confirm deleting items
-StaticPopupDialogs["PASTLOOT_CONFIRM_ITEM_DELETE"] = {
-	text = "%s\nShould PastLoot delete all instances of this item?\nChoice lasts for the remainder of this session.",
-	button1 = "Yes",
-	button2 = "No",
-	OnAccept = function(self, clink)
-		PASTLOOT_CALL_BAG_UPDATE(clink, true)
-	end,
-	OnCancel = function(self, clink)
-		PASTLOOT_CALL_BAG_UPDATE(clink, false)
-	end,
-	timeout = 0,
-	whileDead = true,
-	hideOnEscape = true,
-}
+local PASTLOOT_POPUP = "PASTLOOT_CONFIRM_DELETE_YES_NO"
+
 PastLoot.PASTLOOT_CONFIRMED_ITEM_CHOICES = {}
 function PASTLOOT_CALL_BAG_UPDATE(clink, choice)
 	PastLoot.PASTLOOT_CONFIRMED_ITEM_CHOICES[clink] = choice
@@ -529,6 +516,22 @@ function PastLoot:OnInitialize()
 	-- PanelTemplates_UpdateTabs(PastLoot_TabbedMenuContainer)
 	self.DropDownFrame = CreateFrame("Frame", "PastLoot_DropDownMenu", nil, "UIDropDownMenuTemplate")
 	PastLoot:ResetCache()
+
+	-- used to confirm deleting items
+	StaticPopupDialogs[PASTLOOT_POPUP] = {
+		text = "%s\nShould PastLoot delete all instances of this item?\nChoice lasts for the remainder of this session.",
+		button1 = "Yes",
+		button2 = "No",
+		OnAccept = function(self, clink, ...)
+			PASTLOOT_CALL_BAG_UPDATE(clink, true)
+		end,
+		OnCancel = function(self, clink, reason)
+			if reason == "clicked" then PASTLOOT_CALL_BAG_UPDATE(clink, false) end
+		end,
+		timeout = 0,
+		whileDead = true,
+		hideOnEscape = true,
+	}
 end
 
 local function update_sets()
@@ -646,11 +649,10 @@ local function sort_delete(a, b)
 end
 
 local validationQue = {}
-local deleteTimer
 local function deleteValidation()
 	while #validationQue > 0 do
 		local entry = table.remove(validationQue)
-		if not GetContainerItemGUID(entry.bag, entry.slot) then
+		if entry.guid ~= "" then
 			PastLoot:AddLastRoll(entry.guid, true)
 			PastLoot:Pour(entry.msg)
 		end
@@ -677,7 +679,7 @@ function PastLoot:AddLastRoll(guid, destroyed)
 end
 
 -- used to convert quality number to a word to determine if an item needs a delete confirmation
-local num_to_word = {"Common", "Uncommon", "Rare", "Epic", "Legendary", "Artifact", "Vanity"}
+local num_to_word = { "Common", "Uncommon", "Rare", "Epic", "Legendary", "Artifact", "Vanity" }
 
 -- Bucket Event to handle updating the item cache
 function PastLoot:UpdateBags(...)
@@ -749,8 +751,8 @@ function PastLoot:UpdateBags(...)
 	local todelete = self.db.profile.KeepOpen - freespace
 	while #deletecache > 0 and (todelete > 0 or deletecache[1].value == 0) do
 		local citem = table.remove(deletecache, 1)
-		if citem.guid == GetContainerItemGUID(citem.bag, citem.slot) then
-			if citem.rarity == 0 or PastLoot.PASTLOOT_CONFIRMED_ITEM_CHOICES[citem.clink] == true or self.db.profile["Delete"..num_to_word[citem.rarity]] then
+		if citem.guid ~= "" and citem.guid == GetContainerItemGUID(citem.bag, citem.slot) then
+			if citem.rarity == 0 or PastLoot.PASTLOOT_CONFIRMED_ITEM_CHOICES[citem.clink] == true or self.db.profile["Delete" .. num_to_word[citem.rarity]] then
 				PickupContainerItem(citem.bag, citem.slot)
 				DeleteCursorItem()
 				local StatusMsg = self.db.profile.MessageText.destroy
@@ -763,15 +765,14 @@ function PastLoot:UpdateBags(...)
 				validationEntry.msg = StatusMsg
 				validationEntry.guid = citem.guid
 				table.insert(validationQue, validationEntry)
-			elseif PastLoot.PASTLOOT_CONFIRMED_ITEM_CHOICES[citem.clink] == nil then
-				local dialog = StaticPopup_Show("PASTLOOT_CONFIRM_ITEM_DELETE", citem.clink)
+				Timer.NewTimer(.5, deleteValidation)
+				todelete = todelete - 1
+			elseif not StaticPopup_Visible(PASTLOOT_POPUP) and PastLoot.PASTLOOT_CONFIRMED_ITEM_CHOICES[citem.clink] == nil then
+				local dialog = StaticPopup_Show(PASTLOOT_POPUP, citem.clink)
 				if dialog then
 					dialog.data = citem.clink
 				end
-				return
 			end
-			deleteTimer = Timer.NewTimer(.5, deleteValidation)
-			todelete = todelete - 1
 		end
 	end
 end
